==> 1. Assessing.ipynb <==
# Assessing
Use the space below to explore `all_alpha_08.csv` and `all_alpha_18.csv` to answer the quiz questions below.

import pandas as pd
import numpy as np

df1 =pd.read_csv('all_alpha_08.csv')
df2 =pd.read_csv('all_alpha_18.csv')

df1.info()

sum(df1.duplicated()) # number of duplicate rows

df1.isnull().sum(axis=0)  # number of rows having missing values

df2.info()

sum(df2.duplicated())

df2.isnull().sum(axis=0)

df1.nunique()  # non-unique values in columns

df2.nunique()  # non-unique values in columns
